+++ src/commands/docker.command.ts
@@ -12,9 +12,6 @@ export class DockerCommand extends CommandBase {
         super(program)
     }
     override init(): void {
-
-        
-
         const buildCommand = new Command('build')
             .description('Used to build an image.')
             .option('--push', 'Push to repo immediately after build.')
@@ -37,17 +34,17 @@ export class DockerCommand extends CommandBase {
         let dockerFileName: string = projectSettings.dockerFile || 'dockerfile'
 
         try {
-            super.print('Looking for dockerfile!');
+            this.print('Looking for dockerfile!');
             dockerFile = readFileSync(path.join(process.cwd(), dockerFileName), { encoding: 'utf-8' });
         }
         catch (e) {
-            super.print('Unable to locate dockerfile!', COLORS.RED);
+            this.print('Unable to locate dockerfile!', COLORS.RED);
             return;
         }
 
         //Check for labels & Add if needed
         if (dockerFile.indexOf('LABEL') < 0) {
-            super.print('This dockerfile is missing labels! Adding from settings!')
+            this.print('This dockerfile is missing labels! Adding from settings!')
             let hosts: string = "";
 
             dockerFile += '\nLABEL traefik.enable="true"';
@@ -68,21 +65,21 @@ export class DockerCommand extends CommandBase {
         }
 
         //Build image
-        super.print('Building Image!')
+        this.print('Building Image!')
         try {
             execSync(`docker build -t ghcr.io/arandomdeveloperllc/${projectSettings.imageName} .`, { stdio: 'inherit' })
         } catch (e) {
-            super.print('Error building image: ' + e.message, COLORS.RED);
+            this.print('Error building image: ' + e.message, COLORS.RED);
             return;
         }
 
         //Push option
         if (options.opts().push) {
-            super.print('Pushing Image!')
+            this.print('Pushing Image!')
             try {
                 execSync(`docker push ghcr.io/arandomdeveloperllc/${projectSettings.imageName}:latest`, { stdio: 'inherit' })
             } catch (e) {
-                super.print('Error building image: ' + e.message, COLORS.RED);
+                this.print('Error building image: ' + e.message, COLORS.RED);
                 return;
             }
         }
@@ -90,11 +87,11 @@ export class DockerCommand extends CommandBase {
 
     private pushAction(arg: string, options: any) {
         const projectSettings: ARDCliFile = this.getProjectSettings();
-        super.print('Pushing Image!')
+        this.print('Pushing Image!')
         try {
             execSync(`docker push ghcr.io/arandomdeveloperllc/${projectSettings.imageName}:latest`, { stdio: 'inherit' })
         } catch (e) {
-            super.print('Error building image: ' + e.message, COLORS.RED);
+            this.print('Error building image: ' + e.message, COLORS.RED);
             return;
         }
     }
@@ -103,24 +100,24 @@ export class DockerCommand extends CommandBase {
         let cliFile: string = null;
 
         try {
-            super.print('Looking for a .ard-cli file!');
+            this.print('Looking for a .ard-cli file!');
             cliFile = readFileSync(path.join(process.cwd(), '.ard-cli'), { encoding: 'utf-8' });
         }
         catch (e) {
-            super.print('Unable to find .ard-cli file! Please make sure this file exists!', COLORS.RED)
+            this.print('Unable to find .ard-cli file! Please make sure this file exists!', COLORS.RED)
             return;
         }
 
         const projectSettings: ARDCliFile = JSON.parse(cliFile);
-        projectSettings.imageName = projectSettings.name.toLocaleLowerCase().replace(
+        projectSettings.imageName = projectSettings.imageName.length <= 0 ? projectSettings.name.toLocaleLowerCase().replace(
             / /g,
             ''
-        )
+        ) : projectSettings.imageName;
 
-        super.print('--------------------------------------------')
-        super.print(`Here's what was in the CLI file : ${cliFile}`);
-        super.print(`This is the image name: ${projectSettings.imageName}`)
-        super.print('--------------------------------------------')
+        this.print('--------------------------------------------')
+        this.print(`Here's what was in the CLI file : \n${cliFile}`);
+        this.print(`This is the image name: ${projectSettings.imageName}`)
+        this.print('--------------------------------------------')
 
         return projectSettings;
     }