+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0]
+
+## Added
+
+- Init CLI command to build .ard-cli files
+
+## Fixes
+
+- Fixed the super calls from class inheritance.
+- Some type spacing
+
 ## [1.1.0]
 
 ## Added